Añadida la comprobacion en el add de si no existe previamente el propietario en bbdd y en delete y update se comprueba si existe

+++ src/main/java/DAO/PropietarioDAO.java
@@ -2,31 +2,24 @@ package DAO;
 
 import java.util.List;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
-
 import org.hibernate.Session;
 
 import IDAO.IPropietarioDAO;
 import pojo.Propietario;
-import pojo.Vehiculo;
 import utils.HibernateUtils;
-
 
 public class PropietarioDAO implements IPropietarioDAO {
 
-	
 	// Método para añadir un propietario en la BBDD
 	public void addPropietario(Propietario p) {
 		Session session = HibernateUtils.getSessionFactory().openSession(); // abrimos la conexión
-
-		session.save(p); // declaramos la query de inserción que mete el objeto que recibe (p)
-
-		session.beginTransaction().commit(); // se ejecuta la query y se guardan los cambios
+		// se obtiene el propietario de la bbdd si existe
+		Propietario existente = getPropietarioByDni(p.getDni());
+		// si no existe en la bbdd
+		if (existente == null) {
+			session.save(p); // declaramos la query de inserción que mete el objeto que recibe (p)
+			session.beginTransaction().commit(); // se ejecuta la query y se guardan los cambios
+		}
 		session.close(); // se cierra la conexión
 
 	}
@@ -35,14 +28,16 @@ public class PropietarioDAO implements IPropietarioDAO {
 	public void updatePropietario(Propietario p) {
 		Session session = HibernateUtils.getSessionFactory().openSession(); // abrimos la conexión
 		Propietario updated = getPropietarioByDni(p.getDni()); // creamos un objeto Propietario vacio
-		
-		updated.setDni(p.getDni()); // Le metemos los valores del objeto anterior para que se actualice con un nuevo objeto
-		updated.setNombre(p.getNombre());
-		updated.setApellido(p.getApellido());
-		
-		session.update(p); // declaramos la query de actualización que mete el nuevo objeto
-
-		session.beginTransaction().commit(); // se ejecuta la query y se guardan los cambios
+		if(updated != null) {
+			updated.setDni(p.getDni()); // Le metemos los valores del objeto anterior para que se actualice con un nuevo
+										// objeto
+			updated.setNombre(p.getNombre());
+			updated.setApellido(p.getApellido());
+	
+			session.update(p); // declaramos la query de actualización que mete el nuevo objeto
+	
+			session.beginTransaction().commit(); // se ejecuta la query y se guardan los cambios
+		}
 		session.close(); // se cierra la conexión
 
 	}
@@ -50,9 +45,11 @@ public class PropietarioDAO implements IPropietarioDAO {
 	// Método para eliminar un propietario en la BBDD
 	public void deletePropietario(String dni) {
 		Session session = HibernateUtils.getSessionFactory().openSession();
-		Propietario deleted = new Propietario(dni);
-
-		session.delete(deleted);
+		Propietario deleted = getPropietarioByDni(dni);
+		
+		if(deleted != null) {
+			session.delete(deleted);
+		}
 
 		session.beginTransaction().commit();
 		session.close();
@@ -72,13 +69,13 @@ public class PropietarioDAO implements IPropietarioDAO {
 
 	}
 
-	//Método para obetener de la BBDD un propietario con un dni específico
-    private Propietario getPropietarioByDni(String dni) {
-        Propietario p = null; // se crea un objeto propietario null
-        Session session = HibernateUtils.getSessionFactory().openSession(); // se abre la conexión
-        session.beginTransaction(); // se ejecuta la query para obtener resultados
-        p = session.get(Propietario.class, dni); // metemos el valor resultado filtrando por el dni en un obeto
-        session.close(); // cerramos la conexión
-        return p; // devolvemos el objeto
-    }
+	// Método para obetener de la BBDD un propietario con un dni específico
+	private Propietario getPropietarioByDni(String dni) {
+		Propietario p = null; // se crea un objeto propietario null
+		Session session = HibernateUtils.getSessionFactory().openSession(); // se abre la conexión
+		session.beginTransaction(); // se ejecuta la query para obtener resultados
+		p = session.get(Propietario.class, dni); // metemos el valor resultado filtrando por el dni en un obeto
+		session.close(); // cerramos la conexión
+		return p; // devolvemos el objeto
+	}
 }
